+++ Commands/general/donate.js
@@ -7,6 +7,8 @@ module.exports = class Donate extends Command {
   get aliases() { return ["patreon", "paypal"]; }
 
   exec(message) {
+    if(!(Array.isArray(this.client.config.donate) && this.client.config.donate[0]))
+      return message.channel.send("The bot owner hasn't supplied a donate link!");
     message.channel.send(`Support development by donating!\n${this.client.util.linkList(this.client.config.donate)}`);
   }